requirements.txt
mss
pillow
pyautogui
python-dotenv
openai
pytest
black
flake8
